Moved over and deleted old account and preference pages
