--- core/templates/core/user_accounts/password_change/password_change_done.html
@@ -1,16 +1,0 @@
-{% extends 'core/user_accounts/account_base.html' %}
-{% load i18n static %}
-
-{% block title %}
-  Squire - Password Change Done
-{% endblock title %}
-
-{% block content %}
-  <h1>Password Change Successful!</h1>
-  <p>Hooray! {% trans 'Your password was changed.' %}</p>
-  <p>
-    <a href="{% url 'core:user_accounts/account' %}" class="greenify btn btn-dark greenify">
-      View Account Information
-    </a>
-  </p>
-{% endblock content %}
