+++ user_interaction/account_pages/config.py
@@ -2,12 +2,12 @@ from django.urls import path
 
 from user_interaction.config import AccountConfig
 
-from .views import SiteAccountView
+from .views import SiteAccountView, AccountPasswordChangeView, AccountLayoutPreferencesUpdateView
 
 
 class TestAccountConfig(AccountConfig):
     url_keyword = 'site'
-    tab_select_keyword = 'tab_member_info'
+    tab_select_keyword = 'tab_account_info'
     name = 'Account'
     url_name = 'site_account'
 
@@ -15,4 +15,6 @@ class TestAccountConfig(AccountConfig):
         """ Builds a list of urls """
         return [
             path('', SiteAccountView.as_view(), name='site_account'),
+            path('change-password/', AccountPasswordChangeView.as_view(), name='password_change'),
+            path('change-layout/', AccountLayoutPreferencesUpdateView.as_view(), name='layout_change'),
         ]
