+++ user_interaction/templates/user_interaction/account_pages/password_change_form.html
@@ -3,7 +3,7 @@
 {% load generic_field %}
 
 {% block title %}
-  Squire - Password Change
+    Squire - Password change
 {% endblock title %}
 
 {% block content %}
