--- user_interaction/templates/user_interaction/site_account_page.html
@@ -1,20 +1,0 @@
-
-{% extends 'core/user_accounts/account_base.html' %}
-
-{% block title %}
-    Squire - Membership Data
-{% endblock title %}
-
-{% block og-title %}
-    View Knights Membership Information
-{% endblock og-title %}
-{% block og-description %}
-    View your Knights' membership information here.
-{% endblock og-description %}
-
-{% block content %}
-    <h1>Your account</h1>
-    This page will contain information about the account you use to access Squire.
-
-    It is still on the todo list though
-{% endblock content %}
